+++ packages/designer/src/components/hooks/useDataSources.ts
@@ -5,11 +5,12 @@ export function useDataSources() {
   const { engine, project } = useProject();
 
   const apis = computed(() => project.value?.apis || []);
-
+  const meta = computed(() => project.value?.meta || []);
 
   return {
     engine,
     project,
-    apis
+    apis,
+    meta
   };
 }