+++ packages/designer/src/components/hooks/useDeps.ts
@@ -5,9 +5,11 @@ export function useDeps() {
   const engine = useEngine();
   const dependencies = computed(() => engine.project.value?.dependencies ?? []);
   const apis = computed(() => engine.project.value?.apis ?? []);
+  const meta = computed(() => engine.project.value?.meta ?? []);
   return {
     engine,
     dependencies,
-    apis
+    apis,
+    meta
   };
 }