+++ packages/renderer/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.27](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.8.26...@vtj/renderer@0.8.27) (2024-05-08)
+
+
+### Features
+
+* ✨ 支持metaQuery ([94c2879](https://gitee.com/newgateway/vtj/commits/94c287930d3ae9bafab7419673b4ec5c4fc5c73a))
+
+
+
+
+
+## [0.8.26](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.8.25...@vtj/renderer@0.8.26) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
 ## [0.8.25](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.8.24...@vtj/renderer@0.8.25) (2024-05-07)
 
 **Note:** Version bump only for package @vtj/renderer